requirements.txt
# Orion Backend - Production Dependencies (Python 3.13 compatible)

# Core FastAPI dependencies
fastapi==0.115.9
uvicorn[standard]==0.34.2
pydantic>=2.11.0
pydantic-settings>=2.1.0

# Authentication and Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9

# Database
sqlalchemy==2.0.23
alembic==1.13.1
psycopg[binary]>=3.2.2  # PostgreSQL adapter

# Redis and Caching
redis==5.0.1
hiredis==2.2.3

# Celery for background tasks
celery==5.3.4
flower==2.0.1  # Celery monitoring

# HTTP clients
aiohttp>=3.11.0
httpx>=0.28.0

# Document processing
pypdf==3.17.1
python-docx==1.1.0
beautifulsoup4>=4.12.0
chardet==5.2.0

# AI and Embeddings
google-generativeai==0.3.2
anthropic>=0.50.0
sentence-transformers>=2.2.2
numpy>=1.26,<3.0

# Vector Database
chromadb>=1.0.0

# Web Crawling
Crawl4AI==0.6.2
playwright==1.52.0

# Text Processing
markdown==3.5.1
more-itertools==10.8.0

# Monitoring and Logging
loguru==0.7.2

# Environment and Configuration
python-dotenv>=1.0.0

# Additional utilities
email-validator>=2.1.0
Jinja2>=3.1.2
python-dateutil>=2.9.0
pytz>=2025.2
# Core dependencies that are automatically installed by the above packages
# These are included implicitly and don't need explicit declaration:
# aiofiles, aiohappyeyeballs, aiosignal, anyio, attrs, bcrypt, certifi, cffi
# click, colorama, cryptography, greenlet, h11, httpcore, httptools, idna
# jinja2, markupsafe, multidict, packaging, propcache, pyjwt, sniffio
# starlette, typing_extensions, urllib3, uvloop, watchfiles, yarl, etc.
